splitted 'deleteShareWithUserGroup' method from sharingDialog
